Add initial support for encrypted Windows passwords
The current story around Ansible's management of Windows passwords is
pretty poor. They're poorly protected (in a hostvars variable) and can
be easily and accidentally exposed. Instead, we want to do something
similar to the SSH connection, which is to read in the private key
material only when needed, at connection time, and not expose the
sensitive data at higher levels in Ansible (where it isn't needed).

The goal is to have the WinRM connection plugin do "the right thing"
with as little modifications to the rest of the Ansible code, so it
reuses the existing Ansible magic variables as much as possible.

If you define ansible_private_key_file and ansible_encrypted_password
variables for the host, then this will attempt to decrypt the encrypted
password with the private key specified and, if that succeeds, use that
decrypted value as the password.